# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/pos/catalog": {
      "get": {
        "operationId": "PosController_getCatalog",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Pos"]
      }
    },
    "/pos/transactions": {
      "get": {
        "operationId": "PosController_getTransactions",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Pos"]
      }
    },
    "/pos/orders": {
      "get": {
        "operationId": "PosController_getOrders",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Pos"]
      }
    },
    "/pos/orders/{id}/status": {
      "patch": {
        "operationId": "PosController_updateOrderStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Pos"]
      }
    },
    "/pos/order-line-items/{id}/status": {
      "patch": {
        "operationId": "PosController_updateLineItemStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Pos"]
      }
    },
    "/pos/transactions/bulk": {
      "post": {
        "operationId": "PosController_createTransactionsBulk",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Pos"]
      }
    },
    "/pos/modifier-groups/{id}": {
      "get": {
        "operationId": "PosController_getModifierGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Pos"]
      }
    },
    "/pos/modifier-groups/{id}/options": {
      "get": {
        "operationId": "PosController_getModifierGroupOptions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Pos"]
      }
    },
...
    "/ingestion/upload": {
      "post": {
        "operationId": "IngestionController_handleUpload",
        "parameters": [
          {
            "name": "x-org-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Ingestion"]
      }
    },
    "/ingestion/review/{id}": {
      "get": {
        "operationId": "IngestionController_getReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Ingestion"]
      },
      "patch": {
        "operationId": "IngestionController_updateReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Ingestion"]
      }
    },
    "/ingestion/commit": {
      "post": {
        "operationId": "IngestionController_commitReview",
        "parameters": [
          {
            "name": "x-org-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Ingestion"]
      }
    },
...
  },
  "info": {
    "title": "Sous Tools API",
    "description": "The global API for Sous Tools",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}